requirements.txt
fastapi
python-dotenv
uvicorn
aiohttp